Released version 0.0.3. It supports better override capabilities.
When creating a resource via `makeResource` all methods that already
exist on the class are simply ignored instead of overwritten. This
way it is easier and more natural for the user to provide his own
implementations.

Instead of having to alter the class at a later time like so:

```js
class Pokemon {
  save: () => Promise<Pokemon>;
  remove: () => Promise<Pokemon>;

  static one: (id: number) => Promise<Pokemon>;
  static list: (queryParams: ?Object) => Promise<Array<Pokemon>>;
  static page: (queryParams: ?Object) => Promise<Page<Pokemon>>;
}

makeResource(Pokemon, 'api/pokemon');

Pokemon.one = function(id: number) {
  const p = new Pokemon();
  p.id = 1337;
  return Promise.resolve(p);
};
```

The user can simply change the definition inside of the class like so:

```js
class Pokemon {
  save: () => Promise<Pokemon>;
  remove: () => Promise<Pokemon>;

  static one(id: number): Promise<Pokemon> {
    return get(`${baseUrl}/${id}`, queryParams);
  }

  static list: (queryParams: ?Object) => Promise<Array<Pokemon>>;
  static page: (queryParams: ?Object) => Promise<Page<Pokemon>>;
}

makeResource(Pokemon, 'api/pokemon');
```

This is way more natural since it all happens inside of the class definition.

Updated the documentation to show the new way of providing your own
implementations. Gave an example to show as `static` and `instance`
method.

Also made the start of the documentation better by providing a fully
copy past-able version. This should help new users out a lot.

Also updated to flow 0.49.1.