Updated the documentation.
Made the example of overriding a static method more realistic.

It used to demonstrate how a single pokemon may have more information
than one in a list / page, by defining a SinglePokemon type. This type
contained more fields than the Pokemon type.

This is not what we want to promote because then the `save` and `remove`
instance methods will not work as expected. Because it would try to send
an incomplete Pokemon.

What works better in this case is that the Pokemon class includes all
information of a single pokemon, and that you define another type which
includes less information, and not the other way around.

Now the example defines a PagePokemon type which includes less information.
This way the `save` and `remove` methods will still work, when retrieving
a single pokemon. Also you cannot call `save` and `remove` on a PagePokemon
which is what we also want to express.