The `parseJSON` middleware now ignores empty request bodies (204 responses)
Before this would fail because parsing an empty body would cause an error
when doing `response.json()`.

Now whenever a 204 No Content is detected we do not attempt to convert
it to JSON. The same is true for when the Content-Type is not application/json.

Also added a package-lock.json and updated some dependencies such as flow.