Updated docs: id is now now longer a maybe type, updated parseJson middleware.
The reason why the id was a maybe type is because when creating a new
user via the "new" keyword, the id will always be empty. This happens
when creating a new user via a form. Calling methods which expect an
id will then fail on flow.

However the times that we have a user without an id is an edge case.
Most of the time we have a user in our hands which came from the back-end
and therefore have an id, but flow then required you to check for the
id explicitly, which is good, but very annoying.

So with due consideration we have now decided that the benefits of having
a maybe id which catches rare corner cases, does not weigh against the
amount of boilerplate which needs to be written.