Improved the documentation.
Split up the documentation to form a guide. Also updated a lot of
the code examples to make them shorter / correct. All code examples
now use `ts` as the language.

Also added instructions on how to run `jekyll` with the documentation,
plus added `Gemfile` so it can install the `42bv/fortytwo`. Added
all temporary `jekyll` files to the `.gitignore`.