Improvement: the `one` method now accepts a string or a number.
This way the user is not forced to parse a string to a number when
retrieving the `id` from a url path param which is always a string.

Updated the `2. resource.md` documentation so it matches the `Resource`
class exactly. Also fixed some errors in the type of `QueryParam` in the
doc comments.

Also fixed some audit warnings.

Closes #24.