feature: download file
There is no easy way to facilitate a file download. You have to download
the file using a request, create a link and click that link
programmatically to send the file to the browser. I'd like to see that
functionality in this library to have a robust and consistent solution.

Added downloadFile function to provide file downloads to the browser.

Closes #61