improvement: dependency responsibility
We have some hard dependencies that need to be updated every time there
is an update to one of those libraries. That causes a lot of overhead,
while they are not really hard dependencies. This library should work
fine with any of the patch versions of those libraries and the projects
using this library should be responsible for keeping that library
up-to-date.

Changed dependencies to peerDependencies to move responsibility to users
of this library.