feature: page mapper
Request results can be mapped to a specific type using a mapper, but the
page results require the content property to be mapped. This requires
the same closure used on every pageable request, where only the mapper
might be different. This closure could be provided by this library to
decrease the project code size and complexity.

Added mapPage function to be passed to the then on requests to map the
page using the given mapper function.